Add ability to arbitrary item fields as LIN value
If an order record matches an existing record in Koha, that record is thrown
away and the item(s) are added to the existing record.

This means that the preferred ISBN ( or other identifier ) is thrown away as
well and the identifier from the existing record may or may not be the preferred
( or even acceptable ) identifier.

If the vendor can send the preferred id in a order record field, and we map that order
record field in MarcItemFieldsToOrder, we can have the id survive as an item field.
The field itemnotes_nonpublic is a good one for this.

Then, in the new 'LIN from item field' feature, just set the field name
and the qualifier to use. Now the plugin will use that field as the LIN id!

NOTE: This patch assume the id will alwyas be an ISBN. It looks for an existing
ISBN that matches in the record. If it finds none, it will ad the id as an ISBN
to the record. A future development could make this behavior optional and/or allow
the field and subfield used to be customizable.