Skip non-matching san/suffix combos
This option is to help map invoices to the correct EDI Vendor Record, when using multiple Vendor Recors with the same SAN but different suffixes/ordering account numbers.
Enabling it will skip processing the invoice unless the current EDI Vendor Record's SAN ( from the database ) matches the SAN/Suffix combo from the INVOICE.
That way later on, when the correct account is running invoices, it will process it.
Assumes the vendor SAN is in the first NAD+SU, and the Suffix/Account Number is in a second NAD+SU with a 91 qualifier.
NOTE: If vendor sends multiple INVOICES per message, they *must* all be from the same suffix/account number for this to function correctly.