Add ability to map outgoing GIR subfield values to other values
If it is necessary to convert the value of a GIR field to something else, it can be done here. For example, if the branchcodes in your vendor's profiles differ from the branchcodes in Koha, they can be automatically converted. e.g.

LLO:
   BRANCH_A: BranchA

would convert any LLO value of "BRANCH_A" to "BranchA" in the EDI Order message.